switch to stl data strcuture

+++ src/Core/Scene/Scene.cpp
@@ -1,8 +1,8 @@
-#include "Core/Scene/Scene.hpp"
-#include <typeinfo>
+#include <Core/Scene/Scene.hpp>
+#include <Core/Debug/profiler.hpp>
 
 namespace Raindrop::Core::Scene{
-	RAINDROP_API Scene::Scene(Memory::Allocator& allocator, usize capacity) : _allocator{allocator}, _nameToID{allocator}, _freeComponentIDs{allocator}, _entityManager{allocator, capacity}, _entitySignatures{allocator, capacity}, _systemManager{allocator}{
+	RAINDROP_API Scene::Scene(usize capacity) : _entityManager{capacity}, _entitySignatures{capacity}{
 		RAINDROP_profile_function();
 
 		_capacity = capacity;
@@ -18,22 +18,6 @@ namespace Raindrop::Core::Scene{
 
 	RAINDROP_API Scene::~Scene(){
 		RAINDROP_profile_function();
-
-		deleteSystems();
-		
-		for (usize i=0; i<MAX_COMPONENT_COUNT; i++){
-			auto& manager = _componentManagers[i];
-			if (manager) Memory::deallocateDelete(_allocator, *manager);
-		}
-	}
-
-	RAINDROP_API void Scene::deleteSystems(){
-		RAINDROP_profile_function();
-		auto it = _systemManager._systems.front();
-		while (it){
-			Memory::deallocateDelete(_allocator, *it->_system);
-			it++;
-		}
 	}
 
 	RAINDROP_API ID32 Scene::createEntity(){
@@ -45,70 +29,64 @@ namespace Raindrop::Core::Scene{
 
 	RAINDROP_API void Scene::destroyEntity(ID32 entity){
 		RAINDROP_profile_function();
-		RAINDROP_assert(isIDValid(entity));
+		if (!isIDValid(entity)) throw std::invalid_argument("cannot destroy an invalid entity");
+
 		_entityManager.destroy(entity);
 		Signature signature = _entitySignatures.get(entity);
 		_systemManager.entityRemoved(entity, signature);
 	}
 
-	RAINDROP_API bool Scene::exists(ID32 entity) const{
-		RAINDROP_profile_function();
-		return _entityManager.exists(entity);
-	}
-
 
 	RAINDROP_API void Scene::registerComponent(const char* name, usize typeSize, usize typeAlignement){
 		RAINDROP_profile_function();
-		RAINDROP_assert(!_nameToID.has(name) && "cannot register an already registred component");
+		if (_nameToID.find(name) != _nameToID.end()) throw std::invalid_argument("cannot register an already registred component");
 
-		ID32 id = *_freeComponentIDs.front();
+		ID32 id = _freeComponentIDs.front();
 		_freeComponentIDs.pop();
 
-		_nameToID.insert(name, id);
-		
-		_componentManagers[id] = Memory::allocateNew<ComponentManager>(_allocator, _allocator, typeSize, typeAlignement, _capacity);
+		_nameToID.insert({name, id});
+		_componentManagers[id] = std::make_unique<ComponentManager>(typeSize, typeAlignement, _capacity);
 	}
 
 	RAINDROP_API void Scene::removeComponent(const char* name){
 		RAINDROP_profile_function();
-		RAINDROP_assert(_nameToID.has(name) && "cannot remove a non registred component");
+		if (_nameToID.find(name) == _nameToID.end()) throw std::invalid_argument("cannot remove a non registred component");
 
-		const ID32& id = _nameToID.lookup(name);
-		Memory::deallocateDelete(_allocator, _componentManagers[id]);
-		_componentManagers[id] = nullptr;
+		ID32& id = _nameToID[name];
+		_componentManagers[id].release();
 
 		_freeComponentIDs.push(id);
-		_nameToID.remove(name);
+		_nameToID.erase(name);
 	}
 
 	RAINDROP_API void* Scene::getComponent(ID32 entity, ID32 componentID){
 		RAINDROP_profile_function();
-		RAINDROP_assert(isIDValid(componentID));
-		RAINDROP_assert(isIDValid(entity));
-		RAINDROP_assert(hasComponent(entity, componentID));
+		if (isIDValid(componentID)) throw std::invalid_argument("invalid component ID");
+		if (isIDValid(entity)) throw std::invalid_argument("invalid entity ID");
+		if (!hasComponent(entity, componentID)) throw std::runtime_error("the entity does not have this component");
 
 		return _componentManagers[componentID]->get(entity);
 	}
 
 	RAINDROP_API void Scene::setComponent(ID32 entity, ID32 componentID, void* component){
 		RAINDROP_profile_function();
-		RAINDROP_assert(isIDValid(componentID));
-		RAINDROP_assert(isIDValid(entity));
+		if (isIDValid(componentID)) throw std::invalid_argument("invalid component ID");
+		if (isIDValid(entity)) throw std::invalid_argument("invalid entity ID");
 		_componentManagers[componentID]->set(entity, component);
 	}
 
 	RAINDROP_API bool Scene::hasComponent(ID32 entity, ID32 componentID) const{
 		RAINDROP_profile_function();
-		RAINDROP_assert(isIDValid(componentID));
-		RAINDROP_assert(isIDValid(entity));
+		if (isIDValid(componentID)) throw std::invalid_argument("invalid component ID");
+		if (isIDValid(entity)) throw std::invalid_argument("invalid entity ID");
 
 		return _entitySignatures.getBit(entity, componentID);
 	}
 
 	RAINDROP_API void Scene::addComponent(ID32 entity, ID32 componentID, void* component){
 		RAINDROP_profile_function();
-		RAINDROP_assert(isIDValid(componentID));
-		RAINDROP_assert(isIDValid(entity));
+		if (isIDValid(componentID)) throw std::invalid_argument("invalid component ID");
+		if (isIDValid(entity)) throw std::invalid_argument("invalid entity ID");
 
 		if (component){
 			_componentManagers[componentID]->set(entity, component);
@@ -123,7 +101,9 @@ namespace Raindrop::Core::Scene{
 
 	RAINDROP_API void Scene::eraseComponent(ID32 entity, ID32 componentID){
 		RAINDROP_profile_function();
-		RAINDROP_assert(isIDValid(componentID));
+		if (isIDValid(componentID)) throw std::invalid_argument("invalid component ID");
+		if (isIDValid(entity)) throw std::invalid_argument("invalid entity ID");
+
 		Signature oldSignature = _entitySignatures.get(entity);
 		_entitySignatures.setBit(entity, componentID, false);
 		Signature newSignature = _entitySignatures.get(entity);
@@ -131,19 +111,19 @@ namespace Raindrop::Core::Scene{
 		_systemManager.entitySignatureUpdate(entity, oldSignature, newSignature);
 	}
 
-	RAINDROP_API void Scene::pushSystem(System* system, Signature signature){
+	RAINDROP_API void Scene::pushSystem(std::shared_ptr<System> system, Signature signature){
 		RAINDROP_profile_function();
 		_systemManager.pushSystem(system, signature);
 	}
 
-	RAINDROP_API void Scene::popSystem(System* system){
+	RAINDROP_API void Scene::popSystem(std::shared_ptr<System> system){
 		RAINDROP_profile_function();
 		_systemManager.popSystem(system);
 	}
 
 	RAINDROP_API Signature Scene::getComponentSignature(ID32 componentID) const{
 		Signature signature{};
-		RAINDROP_assert(isIDValid(componentID));
+		if (isIDValid(componentID)) throw std::invalid_argument("invalid component ID");
 		signature.set(componentID, true);
 		return signature;
 	}
@@ -153,7 +133,7 @@ namespace Raindrop::Core::Scene{
 	}
 
 	RAINDROP_API ID32 Scene::getComponentID(const char* name) const{
-		return _nameToID.lookup(name);
+		return _nameToID.find(name) != _nameToID.end();
 	}
 
 	RAINDROP_API bool Scene::isIDValid(ID32 id){
